+++ function/SnappyCodec.go
@@ -43,12 +43,12 @@ const (
 )
 
 type SnappyCodec struct {
-	buffer []int
+	buffer []int32
 }
 
 func NewSnappyCodec() (*SnappyCodec, error) {
 	this := new(SnappyCodec)
-	this.buffer = make([]int, MAX_TABLE_SIZE)
+	this.buffer = make([]int32, MAX_TABLE_SIZE)
 	return this, nil
 }
 
@@ -200,8 +200,8 @@ func (this *SnappyCodec) Forward(src, dst []byte) (uint, uint, error) {
 	for srcIdx < ends {
 		// Update the hash table
 		h := (binary.LittleEndian.Uint32(src[srcIdx:srcIdx+4]) * SNAPPY_HASH_SEED) >> shift
-		t := table[h] // The last position with the same hash as srcIdx
-		table[h] = srcIdx
+		t := int(table[h]) // The last position with the same hash as srcIdx
+		table[h] = int32(srcIdx)
 
 		// If t is invalid or src[srcIdx:srcIdx+4] differs from src[t:t+4], accumulate a literal byte
 		if (t == 0) || (srcIdx-t >= MAX_OFFSET) || (kanzi.DifferentInts(src[srcIdx:srcIdx+4], src[t:t+4])) {
